+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 0.3.0 (2021-08-19)
+
+### ⚠ BREAKING CHANGES
+
+- **环境搭建基本完成:** 环境搭建基本完成
+
+### Features
+
+- 测试 11 ([fed1bea](https://github.com/crazyfdf/nocode/commit/fed1bea2568e75ac4b198bb9c519aac345414eb6))
+- **处理 baseform 的 checkbox,switch,radio 问题:** 处理 BaseForm 的 checkbox,switch,radio 问题 ([e243f89](https://github.com/crazyfdf/nocode/commit/e243f8915817ef7606fc590247f70c74ee20e070))
+- **改变添加，头部导航结构，修复了所有的报错问题:** 改变添加，头部导航结构，修复了所有的报错问题 ([565f50f](https://github.com/crazyfdf/nocode/commit/565f50f46a400d8c12a34ca743394587a8b8b425))
+- **环境搭建基本完成:** 环境搭建基本完成 ([e7f42ff](https://github.com/crazyfdf/nocode/commit/e7f42ff86b673d3fd5e329eb9a4a5174e05bcd88))
+- **加上特效:** 加上特效 ([8b7938e](https://github.com/crazyfdf/nocode/commit/8b7938e332d5d76dab10fd6c3ffcc0bce0361b1d))
+- **使用 usereducer 重构 usestate:** 使用 useReducer 重构 useState ([bad99c9](https://github.com/crazyfdf/nocode/commit/bad99c994f7da967385e6d1cc5996f24143fcd7f))
+- **添加脚手架与添加组件库功能:** 添加脚手架与添加组件库功能 ([7c39817](https://github.com/crazyfdf/nocode/commit/7c3981749c7fa0b45cc3d2e8e2005feebf6ee9aa))
+- **添加删除功能:** 添加删除功能 ([c955999](https://github.com/crazyfdf/nocode/commit/c955999c336b094908962a569d43c54ecc48d7a2))
+- **修改 uctoo-request 兼容 get 传递 data 的错误:** 修改 uctoo-request 兼容 get 传递 data 的错误 ([800b481](https://github.com/crazyfdf/nocode/commit/800b481872e9e1c92a7518a101ea66290082f74b))
+- **选择文件目录:** 选择文件目录 ([98ca35a](https://github.com/crazyfdf/nocode/commit/98ca35ac4299b105707224e2e0b8961a8271afb1))
+- **展示组件:** 展示组件 ([a498076](https://github.com/crazyfdf/nocode/commit/a498076195479fa1da8aead192c82c1178940f46))
+- **支持创建 uctconfig 配置:** 支持创建 uctConfig 配置 ([0a2a6a4](https://github.com/crazyfdf/nocode/commit/0a2a6a4b4d03439a24e67cb5d39219f552c7fe91))
+- **支持页面基本配置:** 支持页面基本配置 ([0719411](https://github.com/crazyfdf/nocode/commit/07194111a56af0d768fbe31f67a2b7adeb1f706d))
+
+### Bug Fixes
+
+- **配置 qiankun 前提交:** 配置 qiankun 前提交 ([70e8e94](https://github.com/crazyfdf/nocode/commit/70e8e9422b69efd574649f8c5cfa4a5acaa708bf))
+- **修复一些报错:** 修复一些报错 ([5005dd6](https://github.com/crazyfdf/nocode/commit/5005dd6dfd4a71ec912bc0a540d84da0a83d7b3f))
+- **修改了一些代码结构:** 修改了一些代码结构 ([44bd2d8](https://github.com/crazyfdf/nocode/commit/44bd2d862323c56b8ae7c022d51db770c15d5840))
+
 ## 0.2.0 (2021-07-21)
 
 ### ⚠ BREAKING CHANGES